+++ audio-workshop/src/main/java/com/hypercube/workshop/audioworkshop/common/pcm/SampleToPCMFunction.java
@@ -4,5 +4,5 @@ import java.nio.ByteBuffer;
 
 @FunctionalInterface
 public interface SampleToPCMFunction {
-    void convert(float[][] normalizedInput, ByteBuffer pcmBuffer, int nbSamples, int nbChannels);
+    void convert(double[][] normalizedInput, ByteBuffer pcmBuffer, int nbSamples, int nbChannels);
 }
